# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: frys 工作流管理系统 API
  description: 企业级轻量化工作流管理系统的 REST API
  version: 1.0.0
  contact:
    email: 1666384464@qq.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: 开发环境
  - url: https://api.frys.dev
    description: 生产环境
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT 访问令牌
  schemas:
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
          description: 请求是否成功
        data:
          type: object
          description: 响应数据
        message:
          type: string
          description: 响应消息
        timestamp:
          type: string
          format: date-time
          description: 响应时间戳
    PaginatedResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            items:
              type: array
              items:
                type: object
            pagination:
              type: object
              properties:
                page:
                  type: integer
                  minimum: 1
                limit:
                  type: integer
                  minimum: 1
                  maximum: 100
                total:
                  type: integer
                  minimum: 0
                totalPages:
                  type: integer
                  minimum: 0
        timestamp:
          type: string
          format: date-time
  parameters:
    pageParam:
      name: page
      in: query
      description: 页码（从1开始）
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    limitParam:
      name: limit
      in: query
      description: 每页数量
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
    searchParam:
      name: search
      in: query
      description: 搜索关键词
      required: false
      schema:
        type: string
security:
  - bearerAuth: []
paths:
...
  /api/v1/auth/register:
    post:
      summary: 用户注册
      description: 注册新用户账户
      tags:
        - 认证
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
                - name
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
                name:
                  type: string
                  minLength: 2
      responses:
        '201':
          description: 注册成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: 请求参数错误
        '409':
          description: 用户已存在
  /api/v1/auth/login:
    post:
      summary: 用户登录
      description: 用户登录获取访问令牌
      tags:
        - 认证
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          description: 认证失败
...
  /api/v1/workflows:
    get:
      summary: 获取工作流列表
      description: 分页获取工作流列表
      tags:
        - 工作流管理
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/pageParam'
        - $ref: '#/components/parameters/limitParam'
        - $ref: '#/components/parameters/searchParam'
        - name: status
          in: query
          description: 工作流状态筛选
          required: false
          schema:
            type: string
            enum: [draft, published, archived]
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'

    post:
      summary: 创建工作流
      description: 创建新的工作流定义
      tags:
        - 工作流管理
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - definition
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 255
                description:
                  type: string
                  maxLength: 1000
                definition:
                  type: object
                  description: 工作流定义JSON
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: 请求参数错误
...